third_party/WebKit/LayoutTests/fast/forms/input-type-change-focusout-expected.txt

CONSOLE WARNING: The 'x-webkit-speech' input field attribute is deprecated. Please use the JavaScript API instead.
This tests that changing an input element's type or speech attribute does not cause focusout event.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS focusoutCalled is false
PASS focusoutCalled is false
PASS focusoutCalled is false
PASS focusoutCalled is false
PASS focusoutCalled is false
PASS focusoutCalled is false
PASS successfullyParsed is true

TEST COMPLETE
